requirements.txt
# Music Class Organizer
google-generativeai
python-dotenv
